+++ src/scripts/components/component.ts
