+++ src/scripts/components/pixi.ts
