+++ src/scripts/entities/create-entity.ts
