+++ src/scripts/entities/entity.ts
