+++ src/scripts/tb/entities/templates-entity.ts
@@ -1,5 +1,5 @@
-import { EntityMap } from "../../shared/entities/create-entity"
-import { ColorSwatch } from '../../shared/util/color-swatch'
+import { EntityMap } from "../../entities/create-entity"
+import { ColorSwatch } from '../../util/color-swatch'
 
 export const dummy: EntityMap = {
     name: 'dummy',
