+++ src/scripts/util/event-manager.ts
