+++ src/scripts/util/local-storage-manager.ts
